requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
SQLAlchemy==2.0.36
pydantic==2.8.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.2.0
reportlab==4.2.2
pdfminer.six==20231228
python-docx==1.1.2
requests==2.32.3
Pillow==10.4.0
apscheduler==3.10.4
python-dotenv==1.0.1
psycopg[binary]==3.1.19
